+++ README.md
@@ -1,15 +1,4 @@
+![logo](src/main/resources/assets/vinurl/icon.png)
+
 ## About
-URL Music Discs is a Minecraft mod that introduces a new custom music disc. This music disc can have a Youtube or cloud-hosted file (Dropbox, Discord CDN, etc.) URL written to it, allowing you to play any song you may desire on a Jukebox! Jam out with your friends, maybe even listen to a podcast, add a whole new layer of socialization to your worlds.
-
-## Video Showcase
-[![IMAGE ALT TEXT HERE](https://img.youtube.com/vi/-UA_nDU72xs/0.jpg)](https://www.youtube.com/watch?v=-UA_nDU72xs)
-
-## Recipes
-![Custom Music Disc Recipe](readme-assets/custom-disc-recipe.png)
-
-## Requirements
-This Mod requires the [Fabric API](https://modrinth.com/mod/fabric-api).
-This Mod automatically downloads FFmpeg as well as YT-DLP in order to download and process content. These files are not ran until you enter a world, and may be manually replaced if you would like by replacing them in the configuration folder.
-
-### Footnote
-Servers may whitelist custom URLs in the configuration file.
+VinURL Mod is an exciting fork of the popular URL Music Discs mod by Hoofer. With the VinURL Music Disc, players can download and play any cloud-hosted audio file directly in Minecraft or download them from YouTube, creating a personalized musical experience like never before.
