+++ gradle.properties
@@ -2,16 +2,16 @@
 org.gradle.jvmargs=-Xmx1G
 org.gradle.parallel=true
 
+# Mod Properties
+mod_version=1.0.0
+maven_group=com.vinurl
+archives_base_name=vinurl
+
 # Fabric Properties
 # check these on https://fabricmc.net/develop
 minecraft_version=1.20.4
-yarn_mappings=1.20.4+build.1
+yarn_mappings=1.20.4+build.3
 loader_version=0.15.7
 
-# Mod Properties
-mod_version=1.2.0
-maven_group=urlmusicdiscs
-archives_base_name=urlmusicdiscs
-
-# Dependencies
-fabric_version=0.96.1+1.20.4
+#Fabric api
+fabric_version=0.96.11+1.20.4
