--- src/client/resources/urlmusicdiscs.client.mixins.json
@@ -1,11 +1,0 @@
-{
-	"required": true,
-	"package": "urlmusicdiscs.mixin.client",
-	"compatibilityLevel": "JAVA_17",
-	"client": [
-      "SoundLoaderMixin"
-	],
-	"injectors": {
-		"defaultRequire": 1
-	}
-}
