+++ src/client/resources/vinurl.client.mixins.json
@@ -1,0 +1,11 @@
+{
+	"required": true,
+	"package": "com.vinurl.mixin.client",
+	"compatibilityLevel": "JAVA_17",
+	"client": [
+      "SoundLoaderMixin"
+	],
+	"injectors": {
+		"defaultRequire": 1
+	}
+}
