--- src/main/resources/assets/urlmusicdiscs/lang/en_us.json
@@ -1,4 +1,0 @@
-{
-  "item.urlmusicdiscs.custom_record": "Custom Music Disc",
-  "item.urlmusicdiscs.custom_record.desc": "Custom Song"
-}
