--- src/main/resources/assets/urlmusicdiscs/models/item/custom_record.json
@@ -1,6 +1,0 @@
-{
-  "parent": "item/generated",
-  "textures": {
-    "layer0": "urlmusicdiscs:item/custom_record"
-  }
-}
