--- src/main/resources/assets/urlmusicdiscs/sounds.json
@@ -1,10 +1,0 @@
-{
-  "placeholder_sound": {
-    "sounds": [
-      {
-        "name": "urlmusicdiscs:placeholder_sound",
-        "stream": true
-      }
-    ]
-  }
-}
