+++ src/main/resources/assets/vinurl/lang/en_us.json
@@ -1,0 +1,4 @@
+{
+  "item.vinurl.custom_record": "Custom Music Disc",
+  "item.vinurl.custom_record.desc": "Custom Song"
+}
