+++ src/main/resources/assets/vinurl/models/item/custom_record.json
@@ -1,0 +1,6 @@
+{
+  "parent": "item/generated",
+  "textures": {
+    "layer0": "vinurl:item/custom_record"
+  }
+}
