+++ src/main/resources/assets/vinurl/textures/gui/record_input.png
