+++ src/main/resources/data/minecraft/tags/items/music_discs.json
@@ -1,5 +1,5 @@
 {
   "values": [
-    "urlmusicdiscs:custom_record"
+    "vinurl:custom_record"
   ]
 }
