+++ src/main/resources/data/urlmusicdiscs/recipes/custom_record.json
@@ -31,7 +31,7 @@
     }
   ],
   "result": {
-    "item": "urlmusicdiscs:custom_record"
+    "item": "vinurl:custom_record"
   }
 }
