--- src/main/resources/urlmusicdiscs.mixins.json
@@ -1,11 +1,0 @@
-{
-	"required": true,
-	"package": "urlmusicdiscs.mixin",
-	"compatibilityLevel": "JAVA_17",
-	"mixins": [
-		"JukeboxMixin"
-	],
-	"injectors": {
-		"defaultRequire": 1
-	}
-}
